M6 Progress
- Small alignment changes to Login and Register.xml
- Added new strings to string.xml
- Made instances private in most .java files
- Commented out the use of Factory and AccountManager imports where they
were sitting idly.
- Changed "status" to "registrationStatus" within RegistrationActivity,
as well as simplifying the registration success section.
- Changed the hash map instance in AccountManager to accountHM for
clarity, as well as filling in most of its methods for use with it (even
though it is temporary).
- Added the 3-strikes login attempt clause to "getAccount" within
Account Manager. This should be working correctly (locking the
accounts).
- Deleted User.java -- I think this was  an old placeholder for
Account.java
- Heavily updated Main.xml. DO NOT CLICK THE BUTTONS (pictures/"add
item"). They are set to do something on press, but no linked to
anything. It'll crash the app right now. I'll be updating this with
scroll/auto-load soon.
- added placeholder picture (question mark).
- minor syntax errors here and there.