Clean-up and bug fix
- Deleting a user now deletes his items (to avoid the instance of
attempting to view a null account)
- Got rid of excess methods in ItemManager
* "clearToFindAccountActivity is still off a bit. Right now it makes
sure you can't use the back button to get to a user's items after
they've been deleted. However, it also makes it so that after deleting a
user and getting sent to the FindAccount page, pressing back takes you
out of the app (instead of main activity or where ever else). IF you're
going to try to fix this, make sure this sequence works (as it is what
originally caused the bug):
AS AN ADMIN... Find Items -> View Item (not your own) -> "I have
found/lost this item" button -> edit account -> delete account ->
(should get sent to find account) -> PRESS THE BACK BUTTON

If you get sent to the item that took you to the deleted account,
pressing the "i have found..." button will cause an error.